Layout Interface

At its core, each layout algorithm is a mapping

adj_matrix ↦ node_positions

where each algorithm has several parameters. The main goal of the following interface is to keep the separation between parameters and function call. Each Algorithm is implemented as subtype of AbstractLayout.

NetworkLayout.AbstractLayout — Type

abstract type AbstractLayout{Dim,Ptype} end

Supertype for all layouts. Each layout Algorithm <: AbstractLayout needs to implement

layout(algo::Algorithm, adj_matrix)::Vector{Point{Dim,Ptype}}

which takes the adjacency matrix representation of a network and returns a list of node positions. Each Algorithm object holds all of the necessary parameters.

By implementing layout the algorithm also inherits the function-like property

Algorithm(; kwargs...)(adj_matrix) -> node_positions

Therefore, each Algorithm <: AbstractLayout is a functor and can be passed around as a function adj_matrix ↦ node_positions which encapsulates all the parameters. This is handy for plotting libraries such as GraphMakie.jl.

There are some additional guidelines:

All of the parameters should be keyword arguments, i.e. it should be allways possible to call Algorithm() without specifying any parameters.

Algorithms should allways return Vector{Point{dim,Ptype}}. If the type or dimensions can be altered use the keywords dim and Ptype for it.

Some parameters may depend on the specific network (i.e. length of start positions vector). If possible, there should be a fallback option (i.e. truncate the list of start positions if network is to small or append with random values).

Iterative Layouts

Iterative layouts are a specific type of layouts which produce a sequence of positions rather than a single list of positions. Those algorithms are implemented as subtypes of IterativeLayout:

NetworkLayout.IterativeLayout — Type

abstract type IterativeLayout{Dim,Ptype} <: AbstractLayout{Dim,Ptype} end

Supertype for iterative layouts. Instead of implementing layout directly, subtypes Algorithm<:IterativeLayout need to implement the iterator interface

Base.iterate(iter::LayoutIterator{<:Algorithm})
Base.iterate(iter::LayoutIterator{<:Algorithm}, state)

where the iteration item is a Vector{Point{Dim,Ptype}} and the iteration state depends on the algorithm.

By implementing the iterator interface the Algorithm inherits the layout and function-like call

layout(algo::Algorithm, adj_matrix) -> node_postions
Algorithm(; kwargs...)(adj_matrix) -> node_positions

One can instantiate an iterable object LayoutIterator

NetworkLayout.LayoutIterator — Type

LayoutIterator(algorithm::IterativeLayout, adj_matrix)

This type bundles an IterativeLayout with an adjacency matrix to form an iterable object whose items are the node positions.

Example

for p in LayoutIterator(Stress(), adj_matrix)
    # do stuff with positions p
end